attempt to push data

+++ scripts/pushit.sh
@@ -1,21 +1,34 @@
 #!/usr/bin/env bash
 
-repo="${1:-missing}"
-slug="${2:-missing}"
-email="${3:-missing}"
-token="${4:-missing}"
+branch="${1:-gh-pages}"
+repo="${2:-missing}"
+slug="${3:-missing}"
+email="${4:-missing}"
+token="${5:-missing}"
 
-cd pages
-git config --global user.name "${slug}[bot]"
-git config --global user.email "${email}"
-git remote set-url origin "https://${slug}[bot]:${token}@github.com/${repo}.git"
-ls
-git status
-# remove everything in the old site
-git rm -r "*" || echo "nothing to do"
-cp -R ../_site/* .
-touch .nojekyll
-git add . && git commit --amend -m 'deploy'
-git status
-git push --force
-cd 
+if [[ "$branch" == "gh-pages" ]]; then
+  cd pages
+  git config --global user.name "${slug}[bot]"
+  git config --global user.email "${email}"
+  git remote set-url origin "https://${slug}[bot]:${token}@github.com/${repo}.git"
+  ls
+  git status
+  # remove everything in the old site
+  git rm -r "*" || echo "nothing to do"
+  cp -R ../_site/* .
+  touch .nojekyll
+  git add . && git commit --amend -m 'deploy'
+  git status
+  git push --force
+  cd 
+else 
+  cd data
+  git config --global user.name "${slug}[bot]"
+  git config --global user.email "${email}"
+  git remote set-url origin "https://${slug}[bot]:${token}@github.com/${repo}.git"
+  ls
+  git status
+  cp -R ../out/* .
+  git add . && git commit -m 'update data'
+  git push
+fi
